+++ components/common/icons/CollectionsIcon.tsx
@@ -1,0 +1,27 @@
+const CollectionsIcon = ({
+  className,
+}: {
+  readonly className?: string | undefined;
+}) => (
+  <svg
+    className={className}
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+    strokeWidth="1.65"
+  >
+    <g transform="translate(1.1325 0) scale(0.125) translate(-48 -33)">
+      <g
+        transform="translate(0,250) scale(0.1,-0.1)"
+        fill="currentColor"
+        stroke="none"
+      >
+        <path d="M497 2152 c-16 -17 -17 -97 -17 -944 0 -883 1 -926 18 -941 17 -16 91 -17 854 -17 798 0 836 1 852 18 15 17 16 60 14 458 l-3 439 -25 0 -25 0 -3 -427 -2 -428 -805 0 c-443 0 -807 0 -810 0 -3 0 -5 405 -5 900 0 495 2 900 5 900 3 0 368 0 813 0 l807 0 0 -427 0 -428 25 0 25 0 3 436 c2 414 1 437 -17 457 l-18 22 -835 0 c-796 0 -836 -1 -851 -18z" />
+        <path d="M726 1924 c-14 -14 -16 -92 -16 -714 0 -622 2 -700 16 -714 14 -14 84 -16 624 -16 540 0 610 2 624 16 14 14 16 92 16 714 0 622 -2 700 -16 714 -14 14 -84 16 -624 16 -540 0 -610 -2 -624 -16z m1205 -716 c0 -669 0 -673 -21 -676 -18 -3 -20 2 -20 45 0 87 -40 238 -80 304 -19 32 -68 95 -97 124 -14 15 -18 26 -12 32 7 7 9 102 7 254 l-3 244 -33 65 c-66 127 -152 187 -287 197 -163 13 -291 -68 -346 -219 -17 -46 -19 -80 -19 -282 0 -215 -1 -233 -22 -275 -12 -25 -24 -48 -28 -51 -17 -15 -69 -92 -92 -137 -35 -70 -68 -192 -68 -254 0 -46 -2 -50 -21 -47 l-21 3 1 670 c1 369 1 672 1 675 0 3 261 4 580 3 l581 -2 0 -673z m-484 526 c64 -21 102 -48 139 -98 62 -84 68 -110 72 -351 l4 -220 -73 -30 c-40 -16 -76 -31 -81 -33 -4 -2 -8 118 -8 266 0 308 -5 333 -81 393 -36 29 -49 33 -104 33 -69 1 -99 -14 -139 -68 -18 -23 -22 -45 -24 -143 -4 -130 7 -186 44 -226 23 -25 24 -29 10 -54 -8 -16 -43 -47 -78 -70 -67 -44 -91 -55 -64 -27 14 13 16 48 16 226 0 228 3 243 62 319 23 31 94 75 148 93 32 10 112 5 157 -10z m-94 -99 c19 -8 46 -28 61 -45 25 -30 26 -37 26 -144 0 -128 -11 -154 -76 -181 -39 -16 -62 -13 -104 11 -47 28 -60 68 -60 184 0 115 12 147 64 173 39 20 46 20 89 2z m107 -360 c0 -22 -2 -24 -14 -13 -10 9 -11 17 -4 26 15 17 18 15 18 -13z m-56 -288 c-150 -133 -262 -87 -209 85 8 27 25 59 40 73 14 14 25 34 25 44 0 15 9 18 60 18 45 0 68 5 88 20 l27 20 3 -113 3 -114 -37 -33z m-234 118 c0 -8 -7 -15 -15 -15 -9 0 -12 6 -9 15 4 8 10 15 15 15 5 0 9 -7 9 -15z m-90 -29 c0 -2 -7 -6 -15 -10 -8 -3 -15 -1 -15 4 0 6 7 10 15 10 8 0 15 -2 15 -4z m54 -48 c-4 -29 -7 -55 -7 -58 -2 -21 20 -63 44 -86 25 -24 38 -28 82 -28 56 1 127 33 179 82 24 22 31 24 51 13 19 -10 32 -8 90 18 65 29 97 32 97 7 0 -7 -16 -18 -35 -25 -76 -27 -147 -158 -180 -331 l-17 -85 -50 1 c-46 2 -52 5 -64 31 -32 71 -93 103 -193 103 l-69 0 -23 43 c-33 62 -32 203 1 270 13 27 37 56 52 65 15 9 28 20 28 24 0 4 5 8 10 8 7 0 8 -19 4 -52z m581 -68 c3 -5 1 -10 -4 -10 -6 0 -11 5 -11 10 0 6 2 10 4 10 3 0 8 -4 11 -10z m-730 -47 c-2 -10 -4 -27 -4 -37 -1 -15 -4 -17 -12 -9 -6 6 -19 9 -28 5 -11 -4 -8 4 9 25 28 36 40 41 35 16z m753 -20 c28 -40 53 -93 64 -133 22 -84 27 -108 29 -164 l3 -58 -168 -2 -168 -1 6 55 c18 149 96 298 169 319 46 13 44 14 65 -16z m-754 -195 c9 -20 25 -46 37 -57 17 -15 19 -23 10 -39 -6 -12 -11 -31 -11 -44 0 -22 -3 -23 -82 -22 -76 1 -83 3 -80 20 16 117 29 163 73 259 l23 50 7 -65 c3 -36 14 -82 23 -102z m90 -64 c3 -8 2 -12 -4 -9 -6 3 -10 10 -10 16 0 14 7 11 14 -7z m177 -53 c16 -16 29 -33 29 -38 0 -4 -46 -7 -102 -6 l-103 1 4 29 c5 42 7 43 77 43 59 0 68 -3 95 -29z m-404 -57 c-3 -3 -12 -4 -19 -1 -8 3 -5 6 6 6 11 1 17 -2 13 -5z m208 -4 c3 -5 1 -10 -4 -10 -6 0 -11 5 -11 10 0 6 2 10 4 10 3 0 8 -4 11 -10z m265 6 c0 -2 -7 -7 -16 -10 -8 -3 -12 -2 -9 4 6 10 25 14 25 6z m560 0 c0 -3 -4 -8 -10 -11 -5 -3 -10 -1 -10 4 0 6 5 11 10 11 6 0 10 -2 10 -4z" />
+      </g>
+    </g>
+  </svg>
+);
+
+export default CollectionsIcon;
