+++ hooks/useSidebarSections.ts
@@ -1,4 +1,4 @@
-import Squares2X2Icon from "@/components/common/icons/Squares2X2Icon";
+import CollectionsIcon from "@/components/common/icons/CollectionsIcon";
 import type { SidebarSection } from "@/components/navigation/navTypes";
 import { AboutSection } from "@/types/enums";
 import {
@@ -43,7 +43,7 @@ export function useSidebarSections(
       {
         key: "collections",
         name: "Collections",
-        icon: Squares2X2Icon,
+        icon: CollectionsIcon,
         items: [
           { name: "The Memes", href: "/the-memes" },
           { name: "6529 Gradient", href: "/6529-gradient" },
